+++ src/lib/iTin.Utilities/iTin.Utilities.Pdf/iTin.Utilities.Pdf.Writer/ComponentModel/Input/Base/IInput.cs
@@ -1,9 +1,12 @@
 ﻿
+using System;
+using System.Collections.Generic;
 using System.IO;
 
 using iTin.Core.ComponentModel;
 using iTin.Core.ComponentModel.Results;
 
+using iTin.Utilities.Pdf.Design.Text;
 using iTin.Utilities.Pdf.Writer.ComponentModel.Input;
 using iTin.Utilities.Pdf.Writer.ComponentModel.Result.Insert;
 using iTin.Utilities.Pdf.Writer.ComponentModel.Result.Output;
@@ -128,6 +131,27 @@ namespace iTin.Utilities.Pdf.Writer.ComponentModel
         /// </para>
         /// </returns>
         IResult SaveToFile(string outputPath, SaveOptions options = null);
+
+        /// <summary>
+        /// Gets the lines of text for this <see cref="PdfInput"/>, filtered values based on a predicate.
+        /// </summary>
+        /// <param name="predicate">A function to test each element for a condition.</param>
+        /// <returns></returns>
+        /// <exception cref="ArgumentException">If document has no pages</exception>
+        /// <exception cref="ArgumentNullException">If <paramref name="predicate"/> is <see langword="null"/></exception>
+        IEnumerable<PdfTextLine> TextLines(Func<PdfTextLine, bool> predicate);
+
+        /// <summary>
+        /// Gets the lines of text for this <see cref="PdfInput"/>, optionally you can set both the start and end pages and a value indicating whether blank lines are included in the result.
+        /// </summary>
+        /// <param name="fromPage">Defines start page. If a value is not set, it will default to 1</param>
+        /// <param name="toPage">Defines end page. If a value is not set, it will default to total document pages</param>
+        /// <param name="removeEmptyLines">Indicates whether blank lines are included in the result. By default they are not included</param>
+        /// <returns></returns>
+        /// <exception cref="ArgumentException">If document has no pages</exception>
+        /// <exception cref="ArgumentOutOfRangeException">If <paramref name="fromPage"/> is less than one or is greater than the total number of pages of the document</exception>
+        /// <exception cref="ArgumentOutOfRangeException">If <paramref name="toPage"/> is less than one or is greater than the total number of pages of the document</exception>
+        IEnumerable<PdfTextLine> TextLines(int? fromPage = null, int? toPage = null, bool removeEmptyLines = true);
 
         /// <summary>
         /// Convert this input into a stream object.